Introduce MacManagementPolicy settings for VSphereDistributedNetwork
Nested under the defaultPortConfig, this change extends the
VSphereDistributedNetwork CRD to support exposure of the
MacManagementPolicy from the Distributed Virtual Switch VMODL1 API.

In particular, this is to support consumers such as Antrea to identify
if forged transmits and MAC learning are enabled.

See
https://developer.broadcom.com/xapis/vsphere-web-services-api/latest/vim.dvs.VmwareDistributedVirtualSwitch.MacManagementPolicy.html
for more information on the parity of the VMODL1 API definitions and
descriptions.